Change Int parse to Long

+++ Review-10-Threads/Example-1/src/main/java/edu/odu/cs/cs330/examples/CoinFlip.java
@@ -55,7 +55,7 @@ public class CoinFlip {
 
         // # Trial argument validation
         try {
-            numTrials = Integer.parseInt(args[1]);
+            numTrials = Long.parseLong(args[1]);
         }
         catch (NumberFormatException e) {
             numTrials = DEFAULT_NUM_TRIALS;
